main/requirements.txt
Django==2.1.7
django-crispy-forms==1.7.2
Flask==1.0.2
google-cloud-container==0.2.1
kubernetes==9.0.0
mysqlclient==1.4.2
PyMySQL==0.9.3
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
requests==2.21.0
urllib3==1.24.1
google-cloud-storage==1.14.0
